Fix Nova cell search
The output from `nova-manage cell_v2 list_cells --verbose` contains
an extra column, stating whether the cell is enabled or not. This means
that the regex never matches, so existing_cells is always empty.

This fix updates the regex by adding a match group for this field which
may be used in a later change.

Unfortuately the CLI doesn't output in JSON format, which would make
this a lot less messy.

Closes-Bug: #1842460
Change-Id: Ib6400b33785f3ef674bffc9329feb3e33bd3f9a3