kolla_address: database_address for non urls
When bootstaping, and other situations, database_address are
used in non url places.
Since DRY is preferred, this commit reuses api_interface_address
instead of creating database_url and replacing the other calls.

One could also do a "ip" context for kolla_address, which would
take the variable and strip it of any brackets/make sure it's an 4/6
number fields with dots or colons as delimiters.

But we need to be able to distinguish between database_address used
in a URL and one used where Ansible/Services interpret it as an IP
(and thus fail since it has brackets in IPv6's case)